requirements.txt
Django==3.0.6
djangorestframework==3.11.0
django-filter==2.2.0
python-decouple==3.3
dj-database-url==0.5.0
psycopg2==2.7.6.1
gunicorn==20.0.4